Rename _sparse_table_morphology.py
