+++ src/_skimage2/morphology/_sparse_table_morphology.py
