+++ src/_skimage2/morphology/meson.build
@@ -3,7 +3,7 @@ python_sources = [
   '__init__.pyi',
   '_grayscale_operators.py',
   '_footprints.py',
-  '_sparse_table.py',
+  '_sparse_table_morphology.py',
 ]
 
 py3.install_sources(
